refactor(git-utils): refactor structure
