+++ packages/amplication-git-utils/src/providers/git-factory.ts
@@ -2,7 +2,7 @@ import { ILogger } from "@amplication/util/logging";
 import { GitProvider } from "../git-provider.interface";
 import { EnumGitProvider, GitProviderArgs } from "../types";
 import { INVALID_SOURCE_CONTROL_ERROR_MESSAGE } from "./git.constants";
-import { GithubService } from "./github.service";
+import { GithubService } from "./github/github.service";
 import { BitBucketService } from "./bitbucket/bitbucket.service";
 
 export class GitFactory {
