+++ packages/amplication-git-utils/src/providers/github/github.service.ts
@@ -9,8 +9,8 @@ import {
   TreeParameter,
   UpdateFunctionFile,
 } from "octokit-plugin-create-pull-request/dist-types/types";
-import { PaginationLimit } from "../errors/PaginationLimit";
-import { GitProvider } from "../git-provider.interface";
+import { PaginationLimit } from "../../errors/PaginationLimit";
+import { GitProvider } from "../../git-provider.interface";
 import {
   Branch,
   CloneUrlArgs,
@@ -37,10 +37,10 @@ import {
   RemoteGitRepository,
   UpdateFile,
   OAuth2FlowResponse,
-} from "../types";
-import { ConverterUtil } from "../utils/convert-to-number";
-import { NotImplementedError } from "../utils/custom-error";
-import { UNSUPPORTED_GIT_ORGANIZATION_TYPE } from "./git.constants";
+} from "../../types";
+import { ConverterUtil } from "../../utils/convert-to-number";
+import { NotImplementedError } from "../../utils/custom-error";
+import { UNSUPPORTED_GIT_ORGANIZATION_TYPE } from "../git.constants";
 
 const GITHUB_FILE_TYPE = "file";
