+++ packages/amplication-git-utils/tests/providers/git-service-factory.spec.ts
@@ -1,10 +1,11 @@
 import { EnumGitProvider } from "../../src/types";
-import { GithubService } from "../../src/git/github.service";
-import { GitFactory } from "../../src/git/git-factory";
+import { GithubService } from "../../src/providers/github/github.service";
+import { GitFactory } from "../../src/providers/git-factory";
 import { mock } from "jest-mock-extended";
 
 describe.skip("GitFactory", () => {
   const github = mock<GithubService>();
+
   describe("GitFactory.getProvider()", () => {
     it("should return an github service", () => {
       expect(
